added sass, restructured component folder😇
