+++ package.json
@@ -14,6 +14,7 @@
     "firebase": "^7.15.4",
     "gh-pages": "^3.1.0",
     "i": "^0.3.6",
+    "node-sass": "^4.14.1",
     "npm": "^6.14.5",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
@@ -22,7 +23,8 @@
     "react-scripts": "3.4.1",
     "react-sound": "^1.2.0",
     "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0"
+    "redux-thunk": "^2.3.0",
+    "sass-loader": "^10.0.5"
   },
   "scripts": {
     "predeploy": "npm run build",
