+++ public/index.html
@@ -7,8 +7,18 @@
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
-      content="Web site created using create-react-app"
+      content="Moodboard app created using React"
     />
+    <!-- OpenGraph -->
+    <meta property="og:url" content="https://oksanasam.github.io/moody/">
+    <meta property="og:title" content="Moody • Moodboard React app">
+    <meta property=”og:type” content=”website” />
+    <meta property="og:description" content=".">
+    <meta property="og:image" content="../MoodyPixel.png">
+    <!-- Twitter -->
+    <meta name="twitter:card" content="summary">
+    <meta name="twitter:site:id" content="@oksanadev">
+    <meta property="twitter:image" content="https://oksanasam.github.io/moody/MoodyPixel.png">
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/favicon.png" />
     <link href="https://fonts.googleapis.com/css2?family=Fredericka+the+Great&display=swap" rel="stylesheet">
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.13.1/css/all.css" integrity="sha384-xxzQGERXS00kBmZW/6qxqJPyxW3UR0BPsL4c8ILaIWXva5kFi7TxkIIaMiKtqV1Q" crossorigin="anonymous">
