+++ src/components/completeImage/index.js
@@ -1,0 +1,4 @@
+import CompleteImage from './CompleteImage';
+
+export default CompleteImage
+
