+++ src/components/completeQuote/index.js
@@ -1,0 +1,4 @@
+import CompleteQuote from './CompleteQuote';
+
+export default CompleteQuote
+
