+++ src/components/footer/Footer.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 //----------LOCAL FILES--------//
-import logo from '../assets/logo.png';
+import logo from '../../assets/logo.png';
 
 function Footer(props) {
   const handlePrint = () => {
