+++ src/components/incompleteImage/index.js
@@ -1,0 +1,3 @@
+import IncompleteImage from './IncompleteImage';
+
+export default IncompleteImage
