+++ src/components/incompleteQuote/index.js
@@ -1,0 +1,3 @@
+import IncompleteQuote from './IncompleteQuote';
+
+export default IncompleteQuote
