+++ src/components/photoBox/index.js
@@ -1,0 +1,3 @@
+import PhotoBox from './PhotoBox';
+
+export default PhotoBox
