+++ src/components/quoteBox/index.js
@@ -1,0 +1,3 @@
+import QuoteBox from './QuoteBox';
+
+export default QuoteBox
