+++ src/styles/App.scss
@@ -1,0 +1,6 @@
+@import 'partials/_setup';
+@import 'partials/_global';
+@import 'partials/_header';
+@import 'partials/_main';
+@import 'partials/_footer';
+@import 'partials/_media_queries';
