+++ src/styles/partials/_global.scss
@@ -1,0 +1,51 @@
+/* general styles */
+html {
+  font-size: 62.5%;
+}
+
+/* stripping default styles */
+body,
+ul,
+h1,
+h2,
+h3,
+h4,
+h5,
+p {
+  margin: 0;
+  padding: 0;
+}
+
+ul {
+  list-style: none;
+}
+
+a {
+  text-decoration: none;
+}
+
+img {
+  max-width: 100%;
+  display: block;
+}
+
+button {
+  background: none;
+  border: none;
+}
+
+/* wrapper styles */
+.wrapper {
+  max-width: 1600px;
+  width: 90%;
+  margin: 0 auto;
+  min-width: 320px;
+}
+
+/* general body styles */
+body {
+  font-family: 'Cutive Mono', monospace;
+  font-size: 1.4rem;
+  line-height: 1.6;
+  height: 100vh;
+}
