+++ src/styles/partials/_media_queries.scss
@@ -1,0 +1,24 @@
+/* MEDIA QUERIES */
+@media (max-width: 768px) {
+  .mainGrid {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
+
+@media (max-width: 480px) {
+  .mainGrid {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+  }
+
+  .frame {
+    width: 100%;
+    padding-top: 20px;
+  }
+
+  .imagePlace img {
+    align-self: center;
+  }
+}
