+++ src/styles/partials/_setup.scss
@@ -1,5 +1,3 @@
-/* MY STYLES */
-
 /* SETUP STARTS */
 html {
   line-height: 1.15;
@@ -793,441 +791,3 @@ html {
 }
 
 /* SETUP ENDS */
-
-/* general styles */
-html {
-  font-size: 62.5%;
-}
-
-/* stripping default styles */
-body,
-ul,
-h1,
-h2,
-h3,
-h4,
-h5,
-p {
-  margin: 0;
-  padding: 0;
-}
-
-ul {
-  list-style: none;
-}
-
-a {
-  text-decoration: none;
-}
-
-img {
-  max-width: 100%;
-  display: block;
-}
-
-button {
-  background: none;
-  border: none;
-}
-
-/* wrapper styles */
-.wrapper {
-  max-width: 1600px;
-  width: 90%;
-  margin: 0 auto;
-  min-width: 320px;
-}
-
-/* general body styles */
-body {
-  font-family: 'Cutive Mono', monospace;
-  font-size: 1.4rem;
-  line-height: 1.6;
-  height: 100vh;
-}
-
-/* HEADER STYLES */
-
-.header {
-  height: 10vh;
-  display: flex;
-  align-items: center;
-}
-
-/* Toggling Colors */
-.header.NotToggledClass {
-  background-color: #d0b8b3;
-}
-
-.header.ToggledClass {
-  background-color: #c6d3dc;
-}
-
-.logo {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-h1 {
-  font-family: 'Fredericka the Great', cursive;
-  font-weight: 500;
-  font-size: 4rem;
-}
-
-/* Color Palette */
-.palette {
-  display: flex;
-  justify-content: space-between;
-  width: 60px;
-  height: 20px;
-  background-color: #ffffff;
-  margin-left: 20px;
-  padding: 3px 3px 3px 3px;
-  box-shadow: 0 1px 1px rgba(0, 0, 0, 0.11), 0 2px 2px rgba(0, 0, 0, 0.11),
-    0 4px 4px rgba(0, 0, 0, 0.11), 0 8px 8px rgba(0, 0, 0, 0.11), 0 16px 16px rgba(0, 0, 0, 0.11),
-    0 32px 32px rgba(0, 0, 0, 0.11);
-}
-
-.colorOne,
-.colorTwo,
-.colorThree {
-  background-color: #f0e7de;
-  width: 30%;
-}
-
-.colorOne.NotToggledClass {
-  background-color: #f0e7de;
-}
-
-.colorTwo.NotToggledClass {
-  background-color: #f7caca;
-}
-
-.colorThree.NotToggledClass {
-  background-color: #d0b8b3;
-}
-
-/* Toggling Colors of Color Palette */
-.colorOne.ToggledClass {
-  background-color: #e4e5e7;
-}
-
-.colorTwo.ToggledClass {
-  background-color: #89a7bf;
-}
-
-.colorThree.ToggledClass {
-  background-color: #c6d3dc;
-}
-
-/* nav styles */
-.navBar {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.rightNav {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-/* Pause Button Styles */
-.pauseButton {
-  color: white;
-  font-size: 25px;
-  outline: none;
-  -webkit-transition: 0.4s;
-  transition: 0.4s;
-  margin-top: 5px;
-  /* padding: 0; */
-}
-
-.pauseButton:hover,
-.pauseButton:focus {
-  cursor: pointer;
-  filter: drop-shadow(8px 8px 10px gray);
-}
-
-/* Color Toggle Switch Styles */
-.switch {
-  position: relative;
-  display: inline-block;
-  width: 60px;
-  height: 34px;
-}
-
-.switch input {
-  opacity: 0;
-  width: 0;
-  height: 0;
-}
-
-.slider {
-  position: absolute;
-  cursor: pointer;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: #89a7bf;
-  -webkit-transition: 0.4s;
-  transition: 0.4s;
-  border-radius: 34px;
-}
-
-.slider:before {
-  position: absolute;
-  content: '';
-  height: 26px;
-  width: 26px;
-  left: 4px;
-  bottom: 4px;
-  background-color: white;
-  -webkit-transition: 0.4s;
-  transition: 0.4s;
-  border-radius: 50%;
-}
-
-input:checked + .slider {
-  background-color: #f7caca;
-}
-
-input:focus + .slider,
-input:hover + .slider {
-  box-shadow: 0 1px 0 0 rgba(255, 255, 255, 0.1), 0 15px 20px -10px rgba(0, 0, 0, 0.5);
-}
-
-input:checked + .slider:before {
-  -webkit-transform: translateX(26px);
-  -ms-transform: translateX(26px);
-  transform: translateX(26px);
-}
-
-/* Main Photo Grid Styles */
-.mainContainer {
-  height: fit-content;
-  min-height: 80vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-/* Toggling Colors */
-.mainContainer.NotToggledClass {
-  background-color: #f0e7de;
-}
-
-.mainContainer.ToggledClass {
-  background-color: #e4e5e7;
-}
-
-.mainGrid {
-  width: 100%;
-  padding: 20px 0;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  justify-items: center;
-  column-gap: 10px;
-  row-gap: 10px;
-}
-
-/* Photo Box Styles */
-.frame {
-  padding-top: 20px;
-  display: flex;
-  height: 320px;
-  min-height: fit-content;
-  width: 50%;
-  max-width: 300px;
-  min-width: 200px;
-  background-color: #ffffff;
-  box-shadow: 0 1px 1px rgba(0, 0, 0, 0.11), 0 2px 2px rgba(0, 0, 0, 0.11),
-    0 4px 4px rgba(0, 0, 0, 0.11), 0 8px 8px rgba(0, 0, 0, 0.11), 0 16px 16px rgba(0, 0, 0, 0.11),
-    0 32px 32px rgba(0, 0, 0, 0.11);
-}
-
-.innerFrame {
-  width: 90%;
-  height: 100%;
-  min-height: fit-content;
-  margin: 0 auto;
-  display: flex;
-  flex-direction: column;
-}
-
-.imagePlace {
-  height: 80%;
-  min-height: fit-content;
-  background-color: #f7caca;
-  display: flex;
-  flex-direction: column;
-  align-self: stretch;
-  justify-content: center;
-  align-items: center;
-  position: relative;
-}
-
-.imagePlace.imageEmpty {
-  background-color: #ffffff;
-  justify-content: flex-start;
-  align-items: center;
-}
-
-/* Toggling Colors */
-.imagePlace.NotToggledClass {
-  background-color: #f7caca;
-}
-
-.imagePlace.ToggledClass {
-  background-color: #89a7bf;
-}
-
-.imagePlace img {
-  align-self: flex-start;
-}
-
-.frame:hover,
-.frame:focus {
-  transform: scale(1, 1) rotate(0deg);
-  transition: all 0.4s;
-}
-
-.icons {
-  display: flex;
-  justify-content: space-between;
-  padding: 10px 5px 5px 0;
-}
-
-.leftIcons {
-  display: flex;
-}
-
-.leftIcons i {
-  padding-right: 10px;
-}
-
-/* INPUT STYLES */
-.moodInput {
-  border-radius: 3px;
-  width: 80%;
-  border: 2px solid black;
-  padding: 7px;
-}
-
-.moodInput::placeholder {
-  font-family: 'Cutive Mono', monospace;
-  font-size: 1.8rem;
-}
-
-/* HELPER BUTTON STYLES */
-.closeButton,
-.addButton {
-  transition: all 0.2s ease-in-out;
-  color: #ffffff;
-  opacity: 0.8;
-  cursor: pointer;
-}
-
-.closeButton:hover,
-.closeButton:focus,
-.addButton:hover,
-.addButton:focus {
-  transform: scale(1.3);
-}
-
-.closeButton:hover,
-.closeButton:focus {
-  color: #000;
-}
-
-.addButton {
-  font-size: 5rem;
-  align-self: center;
-}
-
-.closeButton {
-  position: absolute;
-  top: 0;
-  left: 0;
-  font-size: 2rem;
-}
-
-.randomQuote,
-.quoteAuthor {
-  padding: 10px;
-  font-size: 2rem;
-}
-
-/* FOOTER STARTS */
-.footer {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 10vh;
-  text-align: center;
-}
-
-.footerLogo {
-  max-width: 50px;
-}
-
-.footer.NotToggledClass {
-  background-color: #d0b8b3;
-}
-
-.footer.ToggledClass {
-  background-color: #c6d3dc;
-}
-
-.footerFlex {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.socialIcons {
-  display: flex;
-  justify-content: space-between;
-}
-
-.socialIcons i {
-  color: white;
-  padding: 10px;
-  font-size: 2rem;
-  transition: all 0.2s ease-in-out;
-}
-
-.socialIcons i:hover,
-.socialIcons i:focus {
-  transform: scale(1.5);
-}
-
-/* FOOTER ENDS */
-
-/* MEDIA QUERIES */
-@media (max-width: 768px) {
-  .mainGrid {
-    grid-template-columns: repeat(2, 1fr);
-  }
-}
-
-@media (max-width: 480px) {
-  .mainGrid {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-  }
-
-  .frame {
-    width: 100%;
-    padding-top: 20px;
-  }
-
-  .imagePlace img {
-    align-self: center;
-  }
-}
